notification/requirements.txt
Pyrogram==2.0.96
rollbar==0.16.3
pika==1.3.1
python-dotenv==0.21.0
TgCrypto==1.2.5
peewee